refactor(ios): Relocates AppDelegate KMP install methods
