+++ ios/keyman/Keyman/Keyman/AppDelegate.swift
@@ -25,27 +25,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     // We really should validate that it is a .kmp first... but the app doesn't yet
     // process URL links, so it's fine for now.  (Will change with QR code stuff.)
 
-    // .kmp package install, Keyman 10 onwards
-    var destinationUrl = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)[0]
-    destinationUrl.appendPathComponent("\(url.lastPathComponent).zip")
-    do {
-      let fileManager = FileManager.default
-
-      // For now, we'll always allow overwriting.
-      if fileManager.fileExists(atPath: destinationUrl.path) {
-        try fileManager.removeItem(at: destinationUrl)
-      }
-
-      // Throws an error if the destination file already exists, and there's no
-      // built-in override parameter.  Hence, the previous if-block.
-      try fileManager.copyItem(at: url, to: destinationUrl)
-      installAdhocKeyboard(url: destinationUrl)
-      return true
-    } catch {
-      showKMPError(KMPError.copyFiles)
-      log.error(error)
+    guard let destinationUrl = ResourceFileManager.shared.importFile(url) else {
       return false
     }
+
+    ResourceFileManager.shared.installFile(destinationUrl)
+    return true
   }
 
   func application(_ application: UIApplication,
@@ -117,50 +102,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
       }
     }
     return _overlayWindow!
-  }
-
-  public func installAdhocKeyboard(url: URL) {
-    let documentsDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)[0]
-    var destination =  documentsDirectory
-    destination.appendPathComponent("temp/\(url.lastPathComponent)")
-
-    KeymanPackage.extract(fileUrl: url, destination: destination, complete: { kmp in
-      if let kmp = kmp {
-        self.promptAdHocInstall(kmp)
-      } else {
-        self.showKMPError(KMPError.invalidPackage)
-      }
-    })
-  }
-
-  public func showKMPError(_ error: KMPError) {
-    showSimpleAlert(title: "Error", message: error.rawValue)
-  }
-
-  public func showSimpleAlert(title: String, message: String) {
-    let alertController = UIAlertController(title: title, message: message,
-                                            preferredStyle: UIAlertController.Style.alert)
-    alertController.addAction(UIAlertAction(title: "OK",
-                                            style: UIAlertAction.Style.default,
-                                            handler: nil))
-
-    self.window?.rootViewController?.present(alertController, animated: true, completion: nil)
-  }
-
-  public func promptAdHocInstall(_ kmp: KeymanPackage) {
-    let vc = PackageInstallViewController(for: kmp, completionHandler: { error in
-      if let err = error {
-        if let kmpError = err as? KMPError {
-          self.showKMPError(kmpError)
-        }
-      } else {
-        self.showSimpleAlert(title: "Success", message: "Installed successfully.")
-      }
-    })
-
-    let nvc = UINavigationController.init(rootViewController: vc)
-
-    self.window?.rootViewController?.present(nvc, animated: true, completion: nil)
   }
 
   @objc func registerCustomFonts() {
